Use 3 channel UNET with pre, slice and post
(Not too sure about the dice function now, since the output of the network is single channel,
while the input is 3 channels. So far I'm disregarding the other 2 channels when calculating
the dice function, but we'll see.

Seems like I was able to maintain batch size (to 5) without impacting in memory and the decrease
in augmentation is bringing each epoch to 1.5h, which is more palatable. We'll see about the
performance.