Various fixes on the predict step:
First, the predictions are done in order, so now evaluating a scan makes
sense since the slices NEED to be ordered for the candidates to work.

Improvements on the evaluation step, as I am using multiprocessing now,
which should speed execution of the script at least by 2x.

Added export of predicted masks, so I can visually check what is going
on.

Added evaluation with laplacian preprocessing in place.

Added evaluation of masks other than the spherical
ones.